moving to stdc++11 standart (NULL->nullptr), starting to redesign gui

+++ src/game/common/constants.hpp
@@ -237,7 +237,7 @@ enum SIZE
 
 const float INCREDIBLY_MAX_FLOAT = 1e10; 
 
-const int NONE_ID = 0;
+const unsigned long int NONE_ID = 0;
 const int FAIL_ID = 1;
 const int SUCCESS_ID = 2;
 const int PERFORM_ID = 3;
